requirements.txt
fastapi==0.80.0
pymongo==4.3.3
uvicorn==0.18.3
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
werkzeug==2.3.3
python-multipart==0.0.6